beryllium: overlay: drop the override of fast brightness ramp rate
DisplayPowerController only uses the slow rate for auto brightness,
so changing the fast rate does not affect the smoothness of
transitions.

On the contrary, some apps control the screen brightness by writing
to Settings.System.SCREEN_BRIGHTNESS, for example. In this case,
the value is not considered "temporary", so DisplayPowerController
will try to animate to the target value with the fast rate instead
of SCREEN_ANIMATION_RATE_MINIMUM (0.0f). If the fast rate is too
low, the brightness change caused by an app may have huge delays
or do not work at all. Note that the system brightness slider does
not have this issue because BrightnessController uses
setTemporaryBrightness().

Test: launch VLC and slide up/down on the left side of the screen

Change-Id: I1713fbc89acad52bbabeef83c3d7618b54d818f5
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>